+++ examples/sandbox/docker-swarm.tf
@@ -14,14 +14,15 @@ module "docker-swarm" {
   # version = "~>1.2"
   source = "../../"
 
-  name               = var.name
-  vpc_id             = aws_vpc.main.id
-  managers           = var.managers
-  workers            = var.workers
-  cloud_config_extra = data.template_file.cloud-config.rendered
-  instance_type      = var.instance_type
-  daemon_count       = length(aws_eip.daemons)
-  daemon_eip_ids     = aws_eip.daemons.*.id
+  name                      = var.name
+  vpc_id                    = aws_vpc.main.id
+  managers                  = var.managers
+  workers                   = var.workers
+  cloud_config_extra        = data.template_file.cloud-config.rendered
+  instance_type             = var.instance_type
+  daemon_count              = length(aws_eip.daemons)
+  daemon_eip_ids            = aws_eip.daemons.*.id
+  store_join_tokens_as_tags = true
 
   exposed_security_group_ids = [
     aws_security_group.exposed.id,