+++ managers.tf
@@ -51,11 +51,12 @@ resource "aws_instance" "managers" {
   count         = var.managers
   ami           = data.aws_ami.base_ami.id
   instance_type = local.instance_type_manager
-  subnet_id     = aws_subnet.managers[count.index % length(data.aws_availability_zones.azs.*.names)].id
+  subnet_id     = aws_subnet.managers[count.index % length(data.aws_availability_zones.azs)].id
   private_ip = cidrhost(
-    aws_subnet.managers[count.index % length(data.aws_availability_zones.azs.*.names)].cidr_block,
+    aws_subnet.managers[count.index % length(data.aws_availability_zones.azs)].cidr_block,
     10 + count.index,
   )
+  
 
   # workaround as noted by https://github.com/hashicorp/terraform/issues/12453#issuecomment-284273475
   vpc_security_group_ids = split(
@@ -88,6 +89,9 @@ resource "aws_instance" "managers" {
       ami,
       ebs_block_device,
       user_data_base64,
+      subnet_id,
+      private_ip,
+      availability_zone,
       tags["ManagerJoinToken"],
       tags["WorkerJoinToken"],
     ]