+++ workers.tf
@@ -48,9 +48,9 @@ resource "aws_instance" "workers" {
   count         = var.workers
   ami           = data.aws_ami.base_ami.id
   instance_type = local.instance_type_worker
-  subnet_id     = aws_subnet.workers[count.index % length(data.aws_availability_zones.azs.*.names)].id
+  subnet_id     = aws_subnet.workers[count.index % length(data.aws_availability_zones.azs)].id
   private_ip = cidrhost(
-    aws_subnet.workers[count.index % length(data.aws_availability_zones.azs.*.names)].cidr_block,
+    aws_subnet.workers[count.index % length(data.aws_availability_zones.azs)].cidr_block,
     10 + count.index,
   )
 
@@ -79,6 +79,9 @@ resource "aws_instance" "workers" {
       ami,
       ebs_block_device,
       user_data_base64,
+      subnet_id,
+      private_ip,
+      availability_zone,
     ]
   }