+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## 4.0.1
 
 * Allow using VPC Tags to store the join tokens instead of S3.  This is not the default.
+* Allow policy version updates in the terraform IAM policies
+* Compute region name from instance metadata
 
 ## 4.0.0